% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enfusion_log.R
\name{check_enfusion_log_pw}
\alias{check_enfusion_log_pw}
\title{Check If Enfusion is at Password Login Page}
\usage{
check_enfusion_log_pw(enfusion_path, t)
}
\arguments{
\item{enfusion_path}{file path to enfusion.exe}

\item{t}{Optional, time value to filer latest log entries}
}
\description{
Check If Enfusion is at Password Login Page
}
