% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enfusion_log.R
\name{check_enfusion_log_un}
\alias{check_enfusion_log_un}
\title{Check If Enfusion is at Username Login Page}
\usage{
check_enfusion_log_un(enfusion_path, t)
}
\arguments{
\item{enfusion_path}{file path to enfusion.exe}

\item{t}{Optional, time value to filer latest log entries}
}
\description{
Check If Enfusion is at Username Login Page
}
