% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enfusion_log.R
\name{get_enfusion_log}
\alias{get_enfusion_log}
\title{Get Enfusion Log Entries}
\usage{
get_enfusion_log(enfusion_path, t = NULL)
}
\arguments{
\item{enfusion_path}{file path to enfusion.exe}

\item{t}{Optional, time value to filer latest log entries}
}
\description{
Returns Enfusion Log Entries with timestamps
}
