% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enfusion_report.R
\name{get_enfusion_report}
\alias{get_enfusion_report}
\title{Download Enfusion Report Using API from Excel Add-In}
\usage{
get_enfusion_report(reportWebServiceURL)
}
\arguments{
\item{reportWebServiceURL}{The Enfusion Report URL.
Same as the one used when downloading reports in Excel.}
}
\description{
This function downloads Enfusion reports using the same API as
 the Enfusion Excel Add-In. This cicumvents the need to use the REST API
 which is an additional cost the Enfusion License. It requires logging into
 the enfusion application which can be accomplished using the launch
 enfusion function.
}
\examples{
library(enfusion)
enfusion_process <- launch_enfusion("username", "password")
reportWebServiceURL <- "https://webservices.enfusionsystems.com/mobile/rest/reportservice/exportReport?name=test.trb"
get_enfusion_report(reportWebServiceURL, enfusion_process)
}
