% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_enfusion.R
\name{launch_enfusion}
\alias{launch_enfusion}
\title{Launch Enfusion Application}
\usage{
launch_enfusion(
  username,
  password,
  enfusion_path = file.path("C:", "Program Files", "Enfusion")
)
}
\arguments{
\item{username}{Enfusion Username}

\item{password}{Enfusion Password}

\item{enfusion_path}{Path to Enfusion.exe}
}
\description{
Launch Enfusion Integrata using user name and password
}
